requirements.txt
python-telegram-bot
pytest-html
requests
allure-pytest